SJICube Refactor : Work moved to new_sji.py and corrected
